+++ app/components/Calendar.tsx
@@ -1,4 +1,4 @@
-import { ParsedContributions } from '~/components/home/Contributions'
+import { ParsedContributions } from '~/routes/index'
 
 enum DayColor {
   SMALL = 'bg-tealish-400',
@@ -7,17 +7,24 @@ enum DayColor {
   XLARGE = 'bg-tealish-100',
 }
 
+enum DayThreshold {
+  SMALL = 0,
+  MEDIUM = 8,
+  LARGE = 16,
+  XLARGE = 24,
+}
+
 const Day: React.FC<{ value: number }> = ({ value }) => {
   // TODO these thresholds are arbitrary. Could calculate them dynamically somehow
   let color = DayColor.SMALL
-  if (value >= 24) {
+  if (value >= DayThreshold.XLARGE) {
     color = DayColor.XLARGE
-  } else if (value >= 16) {
+  } else if (value >= DayThreshold.LARGE) {
     color = DayColor.LARGE
-  } else if (value >= 8) {
+  } else if (value >= DayThreshold.MEDIUM) {
     color = DayColor.MEDIUM
   }
-  return <div className={`${color} w-5 h-5 m-1`} />
+  return <div className={`${color} w-5 h-5 m-1 rounded-sm`} />
 }
 
 // TODO tooltip on hover
@@ -42,10 +49,10 @@ const Calendar: React.FC<{ contributions: ParsedContributions }> = ({ contributi
     </div>
     <div className="mt-4 flex flex-row items-center md:justify-end">
       <div>Less</div>
-      <Day value={0} />
-      <Day value={8} />
-      <Day value={16} />
-      <Day value={24} />
+      <Day value={DayThreshold.SMALL} />
+      <Day value={DayThreshold.MEDIUM} />
+      <Day value={DayThreshold.LARGE} />
+      <Day value={DayThreshold.XLARGE} />
       <div>More</div>
     </div>
   </div>
