+++ app/routes/index.tsx
@@ -24,11 +24,51 @@ type ApiData = {
   workflow_runs: number
 }
 
+type ParsedWeek = {
+  date: Date
+  month: string
+  week: Week
+}
+
+type ParsedContributions = Array<{
+  month: string
+  weeks: Array<ParsedWeek>
+}>
+
+type Data = Omit<ApiData, 'contributions'> & { contributions: ParsedContributions }
+
 export async function loader() {
   const data = await fetch('https://capdesk-eng-dashboard.s3.eu-west-1.amazonaws.com/data.json').then((res) =>
     res.json()
   )
-  return json(data)
+  const { contributions, ...rest } = data as unknown as ApiData
+
+  const formatMonth = new Intl.DateTimeFormat('en-GB', { month: 'short' }).format
+
+  const parseWeek = (dateInSeconds: string, week: Week) => {
+    const date = new Date(Number(dateInSeconds) * 1000)
+    // TODO parse the days here to add the date of each contribution { date, value } for the tooltip
+    return {
+      date,
+      month: formatMonth(date),
+      week,
+    }
+  }
+
+  const weeks = Object.entries(contributions).map(([dateInSeconds, week]) => parseWeek(dateInSeconds, week))
+
+  const weeksByMonth = weeks.reduce<Record<string, Array<ParsedWeek>>>((acc, curr) => {
+    acc[curr.month] = acc[curr.month] || []
+    acc[curr.month].push(curr)
+    return acc
+  }, {})
+
+  const parsedContributions: ParsedContributions = Object.entries(weeksByMonth).map(([month, weeks]) => ({
+    month,
+    weeks,
+  }))
+
+  return json({ ...rest, contributions: parsedContributions })
 }
 
 export default function Index() {
@@ -77,4 +117,4 @@ export default function Index() {
   )
 }
 
-export type { ApiData, Week }
+export type { Data, ParsedContributions }
